% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htlgmm.R
\name{htlgmm}
\alias{htlgmm}
\title{htlgmm}
\usage{
htlgmm(
  y,
  Z,
  W = NULL,
  study_info = NULL,
  family = "gaussian",
  A = 1,
  penalty_type = "lasso",
  initial_with_type = "ridge",
  beta_initial = NULL,
  remove_penalty_Z = FALSE,
  remove_penalty_W = FALSE,
  lambda = 0,
  ratio = NULL,
  gamma_adaptivelasso = 1/2,
  summary_type = "multi",
  inference = FALSE,
  use_sparseC = FALSE
)
}
\arguments{
\item{y}{The y is the variable of interest.}

\item{Z}{The overlapping features between main study and external study.}

\item{W}{The unmatched features only in main study, the default is NULL.}

\item{study_info}{The summary statistics for Z only from external study,
which can be summarized in the type of multivariate version or univariate version.}

\item{family}{The family is chosen from c("gaussian","binomial"). Linear regression for "gaussian" and logistic regression for "binomial".}

\item{A}{Study specific adjustment factors, in particular, intercept term in logistic regression. The default is intercept term for

G are the features working for adjustment in reduced model, but G is not summarized in summary statistics(input:study_info).}

\item{penalty_type}{The penalty type for htlgmm, chosen from c("lasso","adaptivelasso","ridge"). The default is "lasso".}

\item{initial_with_type}{Get initial estimation for beta using internal data only
by cross validation using penalty regression, chosen from c("ridge","lasso"). The default is "ridge".}

\item{beta_initial}{The initial estimation for beta if a consistent estimator is available.
E.g., one may input htlgmm result as beta_initial for more rounds to refine the final estimation.
The default is NULL, and internal data is used for initial estimation.}

\item{remove_penalty_Z}{Not penalize overlapping features Z if it is TRUE. The default is FALSE.}

\item{remove_penalty_W}{Not penalize unmatched features W if it is TRUE. The default is FALSE.}

\item{lambda}{Without cross validation, fix the lambda. The default is NULL.}

\item{ratio}{The fixed ratio of X for two-lambda strategy. The default is NULL.}

\item{gamma_adaptivelasso}{The gamma for adaptive lasso. Select from c(1/2,1,2). The default is 1/2.}

\item{summary_type}{The summary statistics type, chosen from c("multi","uni"), where the default is "multi".}

\item{inference}{Whether to do post-selection inference, only work for adaptivelasso. The default is FALSE.}

\item{use_sparseC}{Whether to use approximate version of weighting matrix C.
If approximation, use the diagonal of inverse of C(inv_C) to approximate the inv_C. The default is FALSE.
When internal data sample size is limited, use_sparseC = TRUE is recommended.
When internal data sample size is large enough, use_sparseC = FALSE is recommended.}
}
\value{
beta The target coefficient estimation.

fix_lambda The output lambda.

fix_ratio The output ratio.

corrected_pos For post-selection inference, they are the corrected position passing significant level 0.05 after BH adjustment (Benjamini & Hochberg).

nonzero_pos For estimated beta, the nonzero positions.

pval For nonzero_pos, the calculated p values.

nonzero_var For nonzero_pos, the calculated variances.

\itemize{
 \item{beta:} The target coefficient estimation.
 \item{fix_lambda:} The output lambda.
 \item{fix_ratio:} The output ratio.
 \item{corrected_pos:} For post-selection inference, they are the corrected position passing significant level 0.05 after BH adjustment (Benjamini & Hochberg).
 \item{nonzero_pos:} For estimated beta, the nonzero positions.
 \item{pval:} For nonzero_pos, the calculated p values.
 \item{nonzero_var:} For nonzero_pos, the calculated variances.
}
}
\description{
htlgmm
}
\details{
htlgmm: Heterogeneous Transfer Learning via generalized method of moments(GMM).
}
